Added error handling for duplicate usernames and emails
Before, the entire import process would fail if any of the entries in the CSV
file had an email or a username that was already in use. There was also not very
informative error messages: instead there would be a general 500 error shown in
the modal box. Now, we have added in checks for the duplicate values and will
continue past that entry if there are conflicts so the import process as a whole
can proceed.