Required user fields checked before creating an account
Required fields are a configurable setting in the default.php file -- we have
initialized it to be the standard UserFrosting fields. If any of the fields
are not included in the CSV or have no value provided, a banner message will
be created and the entry will be skipped in the import process.